requirements.txt
mlflow==2.10.0
cloudpickle==3.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pyyaml==6.0.1
scikit-learn==1.4.1.post1
scipy==1.12.0
xgboost==2.0.3
lightgbm==4.1.0
hydra-core==1.3.2
omegaconf==2.3.0